rest_api/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.78.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httptools==0.4.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
neo4j==4.4.5
priority==2.0.0
pydantic==1.9.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
wsproto==1.1.0
